Gradius

+++ maladius/game-logic.js
@@ -1,12 +1,9 @@
 // this generates the canvas tag on index.html with width/height attributes
-console.log("2");
-var game = new Phaser.Game(900, 600, Phaser.AUTO, 'phaser-example', { preload: preload, create: create, update: update, render: render });
-console.log("4");
+let game = new Phaser.Game(900, 600, Phaser.AUTO, 'phaser-example', { preload: preload, create: create, update: update, render: render });
 
 function preload() {
 
     // load all sprites
-    console.log("9");
     game.load.image('bullet', 'img/bullet.png');
     game.load.image('enemyBullet', 'img/enemy-bullet.png');
     game.load.spritesheet('invader', 'img/invader32x32x4.png', 32, 32);
@@ -14,84 +11,75 @@ function preload() {
     game.load.spritesheet('kaboom', 'img/explode.png', 128, 128);
     game.load.image('starfield', 'img/starfield.png');
 
-    console.log("17");
 
     // load all sfx and music
     game.load.audio('music1', 'audio/gradius.mp3');
     game.load.audio('sfx_enemy_die', 'audio/enemy-die.wav');
     game.load.audio('sfx_fire', 'audio/fire.wav');
     game.load.audio('sfx_player_hit', 'audio/player-hit.wav');
-    console.log("24");
 }
-
-console.log("27");
-var player;
-var aliens;
-var bullets;
-var bulletTime = 0;
-var cursors;
-var fireButton;
-var explosions;
-var starfield;
-var score = 0;
-var scoreString = '';
-var scoreText;
-var lives;
-var enemyBullet;
-var firingTimer = 0;
-var stateText;
-var livingEnemies = [];
-var music; 
-var sfx_fire;
-var sfx_enemy_die;
-
-console.log("48");
+let player;
+let aliens;
+let bullets;
+let bulletTime = 0;
+let bulletDelay = 200;
+let cursors;
+let fireButton;
+let explosions;
+let starfield;
+let score = 0;
+let scoreString = '';
+let scoreText;
+let lives;
+let enemyBullet;
+let firingTimer = 0;
+let stateText;
+let livingEnemies = [];
+let music; 
+let sfx_fire;
+let sfx_enemy_die;
+let player_win = false;
+let bullet_speed = 120;
+let addEnnemies = false;
+let ennemiesDelay = 0;
+let konamiCode = false;
 
 function create() {
 
-    console.log("52");
     game.physics.startSystem(Phaser.Physics.ARCADE);
-    console.log("54");
 
     music = game.add.audio('music1');
     music.volume = 0.4;
     music.play();
 
-    console.log("60");
 
 	//	Here we set-up our audio sprites
     sfx_fire = game.add.audio('sfx_fire');
     sfx_fire.allowMultiple = false;
 
-    console.log("66");
 
     sfx_player_hit = game.add.audio('sfx_player_hit');
     sfx_player_hit.allowMultiple = true;
 
-    console.log("71");
 
     sfx_enemy_die = game.add.audio('sfx_enemy_die');
     sfx_enemy_die.allowMultiple = true;
 
-    console.log("76");
 
     //  The scrolling starfield background
     starfield = game.add.tileSprite(0, 0, 900, 600, 'starfield');
 
-    console.log("81");
 
     //  The starship
     player = game.add.sprite(100, 200, 'ship');
     player.anchor.setTo(0.5, 0.5);
     game.physics.enable(player, Phaser.Physics.ARCADE);
 
-    console.log("88");
 
     //  Our two animations, moving up and down.
     player.animations.add('up', [3, 4], 2, false);
     player.animations.add('down', [0, 1], 2, false);
 
-    console.log("94");
 
     //  Our bullet group
     bullets = game.add.group();
@@ -103,7 +91,6 @@ function create() {
     bullets.setAll('outOfBoundsKill', true);
     bullets.setAll('checkWorldBounds', true);
 
-    console.log("106");
 
     // The enemy's bullets
     enemyBullets = game.add.group();
@@ -115,87 +102,74 @@ function create() {
     enemyBullets.setAll('outOfBoundsKill', true);
     enemyBullets.setAll('checkWorldBounds', true);
 
-    console.log("118");
 
     //  The bad guys
     aliens = game.add.group();
     aliens.enableBody = true;
     aliens.physicsBodyType = Phaser.Physics.ARCADE;
 
-    console.log("125");
 
     createAliens();
 
-    console.log("129");
 
     //  The score
     scoreString = 'Score: ';
-    scoreText = game.add.text(10, 10, scoreString + score, { font: '124px Arial', fill: '#fff' });
+    scoreText = game.add.text(10, 10, scoreString + score, { font: '72px Consolas', fill: '#fff' });
 
-    console.log("135");
 
     //  Lives
     lives = game.add.group();
-    game.add.text(game.world.width - 100, 10, 'Health: ', { font: '24px Arial', fill: '#fff' });
+    game.add.text(game.world.width - 100, 10, 'Health: ', { font: '24px Consolas', fill: '#fff' });
 
-    console.log("141");
 
     //  Text
-    stateText = game.add.text(game.world.centerX,game.world.centerY,' ', { font: '32px Arial', fill: '#fff' });
+    stateText = game.add.text(game.world.centerX,game.world.centerY,' ', { font: '32px Consolas', fill: '#fff' });
     stateText.anchor.setTo(0.5, 0.5);
     stateText.visible = false;
 
-    console.log("148");
 
-    for (var i = 0; i < 3; i++) {
+    for (let i = 0; i < 3; i++) {
 
-        console.log("152");
 
-        var ship = lives.create(game.world.width - 150 + (60 * i), 60, 'ship');
+
+        let ship = lives.create(game.world.width - 150 + (60 * i), 60, 'ship');
         ship.anchor.setTo(0.5, 0.5);
         ship.angle = 0;
         ship.alpha = 0.4;
 
-        console.log("159");
+
 
     }
 
-    console.log("163");
 
     //  An explosion pool
     explosions = game.add.group();
     explosions.createMultiple(30, 'kaboom');
     explosions.forEach(setupInvader, this);
 
-    console.log("170");
 
     //  And some controls to play the game with
     cursors = game.input.keyboard.createCursorKeys();
     fireButton = game.input.keyboard.addKey(Phaser.Keyboard.SPACEBAR);
     
-    console.log("176");
 
 }
 
-console.log("180");
-
 function update() {
 
-    console.log("184");
 
     //  Scroll the background
     starfield.tilePosition.x -= 3;
 
-    console.log("189");
 
     if (player.alive) {
 
-        console.log("193");
+
 
         //  Reset the player, then check for movement keys
         player.body.velocity.setTo(0, 0);
 
-        console.log("198");
+
 
         if (cursors.left.isDown) {
             player.body.velocity.x = -200;
@@ -204,7 +178,7 @@ function update() {
             player.body.velocity.x = 200;
         }
 
-        console.log("211");
+
 
         // keyboard up/down
         if (cursors.up.isDown) {
@@ -221,95 +195,98 @@ function update() {
         }
 
 
-        console.log("224");
+        // Konami code?
+        if (konamiCode == true) {
+            konamiCode = false;
+
+            //player life up to 30
+            lives.callAll('revive');
+            for (let i = 3; i < 30; i++) {
+                let ship = lives.create(game.world.width - 150 - (60 * (i-3)), 60, 'ship');
+                ship.anchor.setTo(0.5, 0.5);
+                ship.angle = 0;
+                ship.alpha = 0.4;      
+            }
+
+            //reduce player delay
+            bulletDelay = 50;
+        }
+
 
         //  Firing?
         if (fireButton.isDown) {
             fireBullet();
         }
 
-        console.log("231");
+
 
         if (game.time.now > firingTimer) {
             enemyFires();
         }
 
-        console.log("237");
+
 
         //  Run collision
         game.physics.arcade.overlap(bullets, aliens, collisionHandler, null, this);
-        console.log("241");
+
 
         game.physics.arcade.overlap(enemyBullets, player, enemyHitsPlayer, null, this);
-        console.log("244");
+
+        if(addEnnemies == true && ennemiesDelay < game.time.now){
+            createAliens();
+            addEnnemies = false;
+        }
 
     }
 
 }
 
-console.log("250");
-
 function createAliens() {
-    console.log("252");
-    for (var y = 0; y < 3; y++) {
-        for (var x = 0; x < 5; x++) {
-            var alien = aliens.create(x * 48, y * 50, 'invader');
+    for (let y = 0; y < 3; y++) {
+        for (let x = 0; x < 5; x++) {
+            let alien = aliens.create(x * 48, y * 50, 'invader');
             alien.anchor.setTo(0.5, 0.5);
             alien.animations.add('fly', [ 0, 1, 2, 3 ], 20, true);
             alien.play('fly');
             alien.body.moves = false;
         }
     }
-    console.log("262");
     aliens.x = 600;
     aliens.y = 250;
 
-    console.log("266");
 
     //  Start the invaders moving. Notice we're moving the Group they belong to, rather than the invaders directly.
-    var tween = game.add.tween(aliens).to( { x: 400 }, 2000, Phaser.Easing.Linear.None, true, 0, 1000, true);
+    let tween = game.add.tween(aliens).to( { x: 400 }, 2000, Phaser.Easing.Linear.None, true, 0, 1000, true);
 
-    console.log("271");
 
     //  When the tween loops it calls descend
     tween.onLoop.add(descend, this);
 
-    console.log("276");
 }
 
-console.log("280");
-
 function setupInvader (invader) {
-    console.log("280");
     invader.anchor.x = 0.5;
     invader.anchor.y = 0.5;
     invader.animations.add('kaboom');
-    console.log("284");
 }
-
-console.log("290");
 
 function descend() {
     aliens.x -= 10;
 }
 
-console.log("296");
-
 function render() {
 
 }
 
-console.log("302");
-
 function fireBullet() {
-    console.log("296");
-    game.add.audio('sfx_fire');
-    sfx_fire.volume = 0.2;
-    sfx_fire.play();
 
-    console.log("301");
     //  To avoid them being allowed to fire too fast we set a time limit
     if (game.time.now > bulletTime) {
+
+        game.add.audio('sfx_fire');
+        sfx_fire.volume = 0.2;
+        sfx_fire.play();
+    
         //  Grab the first bullet we can from the pool
         bullet = bullets.getFirstExists(false);
 
@@ -317,14 +294,11 @@ function fireBullet() {
             //  And fire it
             bullet.reset(player.x+8, player.y);
             bullet.body.velocity.x = 400;
-            bulletTime = game.time.now + 200;
+            bulletTime = game.time.now + bulletDelay;
         }
     }
-    console.log("314");
 
 }
-
-console.log("327");
 
 function collisionHandler (bullet, alien) {
     //  When a bullet hits an alien we kill them both
@@ -340,7 +314,7 @@ function collisionHandler (bullet, alien) {
     scoreText.text = scoreString + score;
 
     //  And create an explosion :)
-    var explosion = explosions.getFirstExists(false);
+    let explosion = explosions.getFirstExists(false);
     explosion.reset(alien.body.x, alien.body.y);
     explosion.play('kaboom', 30, false, true);
 
@@ -348,17 +322,28 @@ function collisionHandler (bullet, alien) {
         score += 1000;
         scoreText.text = scoreString + score;
 
-        enemyBullets.callAll('kill',this);
-        stateText.text = " You Won!, \n Click to restart...";
-        stateText.visible = true;
-        music.stop();
+        if( bullet_speed < 500 ) {
+            //restart the game but keep the score and increase the bullet speed
 
-        // the "click to restart" handler
-        game.input.onTap.addOnce(restart,this);
+            // high up the bullet speed by 30%
+            bullet_speed = bullet_speed * 1.3;
+
+            //  And bring the aliens back from the dead :)
+            aliens.removeAll();
+            addEnnemies = true;
+            ennemiesDelay = game.time.now + 1500;
+        }
+        else {
+            player_win = true;
+            enemyBullets.callAll('kill',this);
+            stateText.text = " You Won!, \n Click to restart...";
+            stateText.visible = true;
+            music.stop();
+            // the "click to restart" handler
+            game.input.onTap.addOnce(restart,this);
+        }
     }
 }
-
-console.log("361");
 
 function enemyHitsPlayer (player,bullet) {
     game.add.audio('sfx_player_hit');
@@ -367,20 +352,23 @@ function enemyHitsPlayer (player,bullet) {
 
     bullet.kill();
 
+    // create an explosion :)
+    let explosion = explosions.getFirstExists(false);
+    explosion.reset(player.body.x, player.body.y);
+    explosion.play('kaboom', 30, false, true);
+
+    if (player_win == true) return;
+
+    // reduce the life of the player
     live = lives.getFirstAlive();
 
     if (live) {
         live.kill();
     }
 
-    //  And create an explosion :)
-    var explosion = explosions.getFirstExists(false);
-    explosion.reset(player.body.x, player.body.y);
-    explosion.play('kaboom', 30, false, true);
-
     // PLAYER DIES
     // When the player dies
-    if (lives.countLiving() < 1) {
+    if (lives.countLiving() < 1 ) {
         player.kill();
         enemyBullets.callAll('kill');
 
@@ -393,8 +381,6 @@ function enemyHitsPlayer (player,bullet) {
         game.input.onTap.addOnce(restart,this);
     }
 }
-
-console.log("397");
 
 function enemyFires() {
     //  Grab the first bullet we can from the pool
@@ -411,21 +397,19 @@ function enemyFires() {
 
     if (enemyBullet && livingEnemies.length > 0) {
         
-        var random=game.rnd.integerInRange(0,livingEnemies.length-1);
+        let random=game.rnd.integerInRange(0,livingEnemies.length-1);
 
         // randomly select one of them
-        var shooter=livingEnemies[random];
+        let shooter=livingEnemies[random];
         // And fire the bullet from this enemy
         enemyBullet.reset(shooter.body.x, shooter.body.y);
 
-        game.physics.arcade.moveToObject(enemyBullet,player,120);
+        game.physics.arcade.moveToObject(enemyBullet,player,randomPourcentRange(bullet_speed,10));
         firingTimer = game.time.now + 2000;
     }
 
 }
 
-
-console.log("428");
 
 function resetBullet (bullet) {
     //  Called if the bullet goes out of the screen
@@ -433,16 +417,16 @@ function resetBullet (bullet) {
 
 }
 
-console.log("436");
-
 function restart() {
-    console.log("420");
     //  A new level starts
     music.stop();
     music.play();
 
     score = 0;
     scoreText.text = scoreString + score;
+    player_win = false;
+    bullet_speed = 120;
+    bulletDelay = 200;
 
     //resets the life count
     lives.callAll('revive');
@@ -454,7 +438,41 @@ function restart() {
     player.revive();
     //hides the text
     stateText.visible = false;
-    console.log("438");
 }
 
-console.log("460");
+function randomPourcentRange(value, range){
+    let val = value * (1 + 2*range*(Math.random() - 0.5)/100);
+    return val;
+}
+
+
+
+
+/******************************************
+ * 
+ *  KONAMI CODE IN KONAMI CODE EASTEREGG
+ * 
+ ******************************************/
+
+
+// suite de touche du konami code
+const pattern = ['ArrowUp', 'ArrowUp', 'ArrowDown', 'ArrowDown', 'ArrowLeft', 'ArrowRight', 'ArrowLeft', 'ArrowRight', 'b', 'a'];
+// touche actuel
+let current = 0;
+
+let keyHandler = function (event) {
+
+
+    if (event.key === pattern[current]) {
+        current++;
+        if (pattern.length === current) {
+            current = 0;
+            konamiCode=true;
+        }
+    }
+    else {
+        current = 0;
+    }
+};
+
+document.addEventListener('keydown', keyHandler, false);
